isolate activation upgrade fixtures from production services
